Implement cart functionality with CartController, CartService, and CartItemRequestDTO. Update Address and Product models for validation. Add custom exceptions for product and user handling.

+++ src/main/java/com/andreibel/ecomapplication/DTO/ProductResponseDTO.java
@@ -3,11 +3,12 @@ package com.andreibel.ecomapplication.DTO;
 import jakarta.validation.constraints.NotNull;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 /**
  * DTO for {@link com.andreibel.ecomapplication.model.Product}
  */
-public record ProductResponseDTO(Long id, @NotNull String name, @NotNull String description, @NotNull Double price,
+public record ProductResponseDTO(Long id, @NotNull String name, @NotNull String description, @NotNull BigDecimal price,
                                  @NotNull Integer stockQuantity, @NotNull String category, @NotNull String imageUrl,
                                  @NotNull Boolean active) implements Serializable {
 }
